// web/src/generated-api/models/BackendType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BackendType = {
    Chat: 'chat',
    Image: 'image'
} as const;
export type BackendType = typeof BackendType[keyof typeof BackendType];


export function instanceOfBackendType(value: any): boolean {
    return Object.values(BackendType).includes(value);
}

export function BackendTypeFromJSON(json: any): BackendType {
    return BackendTypeFromJSONTyped(json, false);
}

export function BackendTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BackendType {
    return json as BackendType;
}

export function BackendTypeToJSON(value?: BackendType | null): any {
    return value as any;
}