// web/src/generated-api/models/DriverInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DriverInfo
 */
export interface DriverInfo {
    /**
     * 
     * @type {string}
     * @memberof DriverInfo
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof DriverInfo
     */
    url: string;
}

/**
 * Check if a given object implements the DriverInfo interface.
 */
export function instanceOfDriverInfo(value: object): boolean {
    if (!('name' in value)) return false;
    if (!('url' in value)) return false;
    return true;
}

export function DriverInfoFromJSON(json: any): DriverInfo {
    return DriverInfoFromJSONTyped(json, false);
}

export function DriverInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DriverInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'url': json['url'],
    };
}

export function DriverInfoToJSON(value?: DriverInfo | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'url': value['url'],
    };
}