// web/src/generated-api/models/ServiceInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ServiceInfo
 */
export interface ServiceInfo {
    /**
     * 
     * @type {string}
     * @memberof ServiceInfo
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceInfo
     */
    providerId: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceInfo
     */
    model: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceInfo
     */
    name: string;
}

/**
 * Check if a given object implements the ServiceInfo interface.
 */
export function instanceOfServiceInfo(value: object): boolean {
    if (!('id' in value)) return false;
    if (!('providerId' in value)) return false;
    if (!('model' in value)) return false;
    if (!('name' in value)) return false;
    return true;
}

export function ServiceInfoFromJSON(json: any): ServiceInfo {
    return ServiceInfoFromJSONTyped(json, false);
}

export function ServiceInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServiceInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'providerId': json['provider_id'],
        'model': json['model'],
        'name': json['name'],
    };
}

export function ServiceInfoToJSON(value?: ServiceInfo | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'provider_id': value['providerId'],
        'model': value['model'],
        'name': value['name'],
    };
}